※人工智能安全/算法1/Code1/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64


python=3.10.13
pytorch=2.1.0
pytorch-cuda=11.8
tensorboard=2.12.1
torchvision=0.16.0
matplotlib=3.7.2
numpy=1.26.0
